third_party/blink/web_tests/virtual/omt-worker-fetch/external/wpt/workers/baseurl/alpha/worker-in-worker-expected.txt

This is a testharness.js-based test.
FAIL Base URL in workers: new Worker() assert_equals: expected "gamma" but got "beta"
Harness: the test ran to completion.
